requirements/requirements-all.txt
jsonschema
logmuse>=0.2.5
oyaml
psycopg2-binary
pydantic>=1.10.7
sqlmodel>=0.0.8
ubiquerg>=0.6.1
yacman>=0.9.0
